refactor: address review feedback for pager
- Add utils.ts with shared pager utilities
- Fix Windows compatibility: use 'where' instead of 'which'
- Support $PAGER with arguments (e.g., "less -R")
- Add FOOTER_RESERVE_LINES constant (replaces magic number 2)
- Extract commandExists, parsePagerCommand, runPager to utils
