+++ src/shell/pager/index.ts
@@ -5,4 +5,5 @@
 export { LessPager } from './less-pager.js';
 export { MorePager } from './more-pager.js';
 export { renderRowsToLines } from './renderer.js';
+export { commandExists, parsePagerCommand, runPager, FOOTER_RESERVE_LINES } from './utils.js';
 export type { Pager, PagerOptions } from './types.js';
