+++ src/shell/pager/more-pager.ts
@@ -11,10 +11,10 @@
  * Note: When using external pager, standard pager key bindings apply.
  */
 
-import { spawn, spawnSync } from 'child_process';
 import type { Pager, PagerOptions } from './types.js';
 import type { PipelineValue } from '../pipeline-types.js';
 import { renderRowsToLines } from './renderer.js';
+import { commandExists, runPager, FOOTER_RESERVE_LINES } from './utils.js';
 
 export class MorePager implements Pager {
   private options: PagerOptions;
@@ -35,7 +35,7 @@ export class MorePager implements Pager {
 
     // If content fits in one page, just print and return
     const terminalHeight = this.options.height ?? (process.stdout.rows || 24);
-    const pageSize = Math.max(1, terminalHeight - 2); // Reserve space for prompt
+    const pageSize = Math.max(1, terminalHeight - FOOTER_RESERVE_LINES);
 
     if (lines.length <= pageSize) {
       lines.forEach(line => console.log(line));
@@ -62,9 +62,9 @@ export class MorePager implements Pager {
     // -R: interpret ANSI color codes
     // -S: don't wrap long lines (preserves table layout)
     // -X: don't clear screen on exit
-    if (this.commandExists('less')) {
+    if (commandExists('less')) {
       try {
-        await this.runPager('less', ['-ERSX'], content);
+        await runPager('less', ['-ERSX'], content);
         return true;
       } catch {
         // less failed, continue to built-in
@@ -72,37 +72,6 @@ export class MorePager implements Pager {
     }
 
     return false;
-  }
-
-  /**
-   * Check if a command exists
-   */
-  private commandExists(cmd: string): boolean {
-    const result = spawnSync('which', [cmd], { stdio: 'ignore' });
-    return result.status === 0;
-  }
-
-  /**
-   * Run a pager command with content
-   */
-  private runPager(cmd: string, args: string[], content: string): Promise<void> {
-    return new Promise((resolve, reject) => {
-      const pager = spawn(cmd, args, {
-        stdio: ['pipe', 'inherit', 'inherit'],
-      });
-
-      pager.on('error', reject);
-      pager.on('close', (code) => {
-        if (code === 0 || code === null) {
-          resolve();
-        } else {
-          reject(new Error(`Pager exited with code ${code}`));
-        }
-      });
-
-      pager.stdin?.write(content);
-      pager.stdin?.end();
-    });
   }
 
   /**
